feat: improve create template script
Refined the create-template script to introduce additional automation
with two new arguments: "all" and "with-previous". The "all" parameter
facilitates the creation of folders for all tasks from the event's start
to its end, streamlining the process. The "with-previous" parameter
option enables the generation of missing tasks from the current date
to the event's start date.