feat: template create automation (#23)
* feat: improve create template script

Refined the create-template script to introduce additional automation
with two new arguments: "all" and "with-previous". The "all" parameter
facilitates the creation of folders for all tasks from the event's start
to its end, streamlining the process. The "with-previous" parameter
option enables the generation of missing tasks from the current date
to the event's start date.

* feat: add new create-template scripts

* docs: update README.md with new scripts

styles: fix whitespaces

---------

Co-authored-by: Przemek Smyrdek <przemek.smyrdek@gmail.com>